refactor out GetProposal and GetCommittee
